Release prep: Moved RouteSpec to its own namespace
